General: Share the stacked label/value block across modules
InfoEntry, groupInfoEntries and InfoBlock move from the viewer module into
app-common so layers below app-workspace-viewer can use them. InfoEntry gains
a valueMaxLines (default 2, unchanged for every existing call site) and
InfoBlock an optional valueLeading slot in front of the value.

visibleInfoRows and permissionsFitHalfWidth stay in the viewer: they describe
that card's collapse behaviour, not the block itself.
